src/python/backend/requirements.txt
# Web Framework
fastapi>=0.109.2
uvicorn>=0.27.1
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.6.1
email-validator>=2.1.0

# Database
sqlalchemy>=2.0.23
alembic>=1.13.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Cache
redis>=5.0.1
aioredis>=2.0.1
pickle-mixin>=1.0.2

# Monitoring & Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-exporter-jaeger>=1.22.0
psutil>=5.9.8
jaeger-client>=4.8.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
cryptography>=42.0.5
pyjwt>=2.8.0
bcrypt>=4.1.2
ipaddress>=1.0.23

# Configuration
python-dotenv>=1.0.1
pyyaml>=6.0.1
pydantic-settings>=2.1.0

# Logging
structlog>=24.1.0
python-json-logger>=2.0.7
loguru>=0.7.2
sentry-sdk>=1.39.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.27.0
faker>=22.6.0

# Development
black>=24.2.0
flake8>=7.0.0
mypy>=1.7.0
isort>=5.13.0
pre-commit>=3.6.0
pytest-watch>=4.2.0

# Performance Optimization
numpy>=1.26.0
scipy>=1.12.0
line-profiler>=4.1.1
memory-profiler>=0.61.0
